requirements.txt
fastapi
python-multipart
pydantic
uvicorn
ffmpeg-python
openai-whisper
transformers
optimum
accelerate
coqui-tts